# Performance feature: use predefined chars arrays for setCookie() method