Hi! 
Here https://github.com/php-curl-class/php-curl-class/compare/php-curl-class:master...maxvodo:with-benchmarking I wrote some benchmarks to examine: how we can increase performance by method ```Curl::setCookie()``` in ```in_array()``` case. 

Some results from my machine:
```
PHP 5.5.9-1ubuntu4.14 (cli) (built: Oct 28 2015 01:34:46) 
Copyright (c) 1997-2014 The PHP Group
Zend Engine v2.5.0, Copyright (c) 1998-2014 Zend Technologies
    with Zend OPcache v7.0.3, Copyright (c) 1999-2014, by Zend Technologies
    with Xdebug v2.2.3, Copyright (c) 2002-2013, by Derick Rethans

Searching benchmarks in: ./php-curl-class/tests/benchmarks
Loading benchmark from: forCookies.php
Running tests 60000 times.
Testing 1/1 :  call CookiesConsts::Init()

Test                            Time   Time (%)   Memory   Memory (%)   
 call CookiesConsts::Init()   123 ms                 0 B                

Running tests 80000 times.
Testing 8/8 :  predefined array_key_exists RFC6265

Test                                      Time   Time (%)   Memory   Memory (%)   
 predefined array_key_exists RFC6265    450 ms                 0 B                
 predefined array_key_exists RFC2616    454 ms        1 %      0 B                
 predefined in_array RFC6265            508 ms       13 %      0 B                
 predefined in_array RFC2616            513 ms       14 %      0 B                
 in_array RFC2616                       852 ms       89 %      0 B                
 in_array RFC6265                       905 ms      101 %      0 B                
 array_key_exists RFC2616              1155 ms      157 %      0 B                
 array_key_exists RFC6265              1250 ms      178 %      0 B               
```

Using of once initialized arrays it's good variant.