@thebeline 2015-10-31
Just saw the merge go in and, while late to the party, figured I would chime in:

I have a few qualms about this. It adds yet another class to the file (not a standard we follow, but still) and the class and execution is unnecessary.  A class whose only responsibility is to contain two arrays and it is overly complicated. The arrays could remain on the Curl class, and could be coded manually, eliminating the need of a separate class and method calls.