@zachborboa 2015-10-31
@thebeline - Works for me. Make `Curl::RFC2616` and `Curl::RFC6265`. Remove `CurlCookieConst`. Correct? Can you send me a pull request with what you had in mind. Thanks