@maxvodo 2015-11-02
@thebeline, @zachborboa - Encapsulating arrays into Curl class it's good idea anyway. I just wanted to describe thing: is needed predefined arrays.